Refactored to use the common parent pom.
