+++ config/SplunkHttpSink.properties
@@ -17,7 +17,7 @@
 name=splunk-http-sink
 topics=syslog-udp
 tasks.max=1
-connector.class=io.confluent.kafka.connect.splunk.SplunkHttpSinkConnector
+connector.class=com.github.jcustenborder.kafka.connect.splunk.SplunkHttpSinkConnector
 splunk.remote.host=confluent
 splunk.remote.port=8088
 splunk.ssl.enabled=true
