+++ config/SplunkHttpSource.example.properties
@@ -16,6 +16,6 @@
 
 name=splunk-http-source
 tasks.max=1
-connector.class=io.confluent.kafka.connect.splunk.SplunkHttpSourceConnector
+connector.class=com.github.jcustenborder.kafka.connect.splunk.SplunkHttpSourceConnector
 splunk.ssl.key.store.path=
 splunk.ssl.key.store.password=
