+++ config/SplunkHttpSource.properties
@@ -16,7 +16,7 @@
 
 name=splunk-http-source
 tasks.max=1
-connector.class=io.confluent.kafka.connect.splunk.SplunkHttpSourceConnector
+connector.class=com.github.jcustenborder.kafka.connect.splunk.SplunkHttpSourceConnector
 splunk.ssl.key.store.path=/Users/jeremy/keystores/keystore.jks
 splunk.ssl.key.store.password=password
 splunk.collector.index.default=main
