--- src/test/java/io/confluent/kafka/connect/splunk/EventIteratorTest.java
@@ -1,42 +1,0 @@
-/**
- * Copyright © 2016 Jeremy Custenborder (jcustenborder@gmail.com)
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.confluent.kafka.connect.splunk;
-
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.databind.JsonNode;
-import org.junit.Test;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-public class EventIteratorTest {
-  public static InputStream events() {
-    return EventIteratorTest.class.getResourceAsStream("events.json");
-  }
-
-  @Test
-  public void iterate() throws IOException {
-    JsonFactory factory = new JsonFactory();
-    InputStream inputStream = events();
-    EventIterator iterator = EventIterator.create(factory, inputStream);
-
-    while (iterator.hasNext()) {
-      JsonNode jsonNode = iterator.next();
-      System.out.println(jsonNode.toString());
-    }
-  }
-
-}
