+++ src/test/resources/com/github/jcustenborder/kafka/connect/splunk/event.minimal.json
