+++ src/test/resources/com/github/jcustenborder/kafka/connect/splunk/events.json
