+++ src/test/resources/com/github/jcustenborder/kafka/connect/splunk/response.400.json
