+++ src/test/resources/com/github/jcustenborder/kafka/connect/splunk/response.403.json
